+++ Backend/Human Resource Management System/src/main/java/org/tjss/humanresourcemanagementsystem/repository/RoleRepository.java
@@ -10,6 +10,6 @@ import org.tjss.humanresourcemanagementsystem.entity.Role;
 
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
   Optional<Role> findByName(ERole name);
 }
